Move module to its own package
